# Optimize `CalculateOffsetHoursFromLongitude`